# crypto/secp256k1: use libsecp256k1 endomorphism