--- a/crypto/secp256k1/secp256.go
+++ b/crypto/secp256k1/secp256.go
@@ -25,6 +25,7 @@ package secp256k1
 #define USE_FIELD_INV_BUILTIN
 #define USE_SCALAR_8X32
 #define USE_SCALAR_INV_BUILTIN
+#define USE_ENDOMORPHISM
 #define NDEBUG
 #include "./libsecp256k1/src/secp256k1.c"
 #include "./libsecp256k1/src/modules/recovery/main_impl.h"