```
benchmark              old ns/op     new ns/op     delta
BenchmarkRecover-2     208484        170297        -18.32%
```

Endomorphism is disabled by default in libsecp256k1 [1], the reason appears to be concerns that it may be covered by patents [2]. The patents in question appears to be owned by Certicom [3][4].

If this implementation is indeed covered by patents, most likely it would be fine for open-source libraries to make use of it. Moreover, as the code is already in place, it's easy to disable in the event of agitated patent holders.

1. https://github.com/bitcoin-core/secp256k1#implementation-details
2. https://bitcointalk.org/index.php?topic=1740973.0
3. https://www.google.com/patents/US20020044649
4. https://en.wikipedia.org/wiki/ECC_patents