@karalabe 2017-01-19
If it is disabled by default, what guarantees are there for correct functioning? I assume that no project uses it extensively, so it's much less battle-tested.

> If this implementation is indeed covered by patents, most likely it would be fine for open-source libraries to make use of it.

I don't see why that would be the case. Unless explicitly stated, OSS isn't a special case.

> Moreover, as the code is already in place, it's easy to disable in the event of agitated patent holders.

Enabling code willingly while knowing it's covered by patents will not be "easily disabled" if someone becomes "agitated". It is a deliberate patent infringement, that can be rightfully pursued in court.

If you can get a written agreement from them for us to use it (or any such agreement for open source projects in general), we can definitely consider it. Until then, I'd hold this off.