@Gustav-Simonsson 2017-01-19
> If it is disabled by default, what guarantees are there for correct functioning? I assume that no project uses it extensively, so it's much less battle-tested.

It passes all the tests in libsecp256k1, which gives a strong guarantee for correct functioning. I don't know of any live crypto system using this option, but wouldn't necessarily wait for one to use it before using it in go-ethereum. It's worth pointing out that go-ethereum began using libsecp256k1 before it had any official release or usage in live systems, without ever running into issues with the library itself.

That said, I'm not arguing this has to be merged anytime soon. It's critical crypto and it would be good to get more eyes on the endomorphism implementation in libsecp256k1.

> I don't see why that would be the case. Unless explicitly stated, OSS isn't a special case.

Corporations tend to use patents as baseball bats in disputes with competing corporations. They are in general less concerned with not-for-profit and non-competing projects.

> Enabling code willingly while knowing it's covered by patents will not be "easily disabled" if someone becomes "agitated". It is a deliberate patent infringement, that can be rightfully pursued in court.

In some, but not all jurisdictions. In what court do you think a potential patent holder would pursue an infringement in this case? And who would be the defendants(s)?

More importantly, in this case it's not _known_ that this code is in fact covered by patents - it may be but it remains a concern, not a fact. So enabling this code cannot be done while _knowing_ that it is covered by patents.

We shouldn't seek agreement from anyone who's patents _may_ cover this, as the burden of proving whether a patent covers this code is on the potential patent holder, not us.