@fjl 2017-01-20
I agree with @Gustav-Simonsson on the patent issue. The code enabled by USE_ENDOMORPHISM needs to be vetted.