@obscuren 2017-01-22
I agree that patents shouldn't hold this off.

However in the scenario where company X holds a patent and company Y would like to build something on go-ethereum it might become a concern. Also a buyout of company X (e.g. oracle buying sun and fucking everyone over using Java) might be a serious concern.