@fjl 2017-02-22
We will not merge this in the foreseeable future. This PR would just stay open forever and the change is trivial so we won't do it. We can reconsider if/when bitcoind enables this.